Only logs Modders Corner messages

+++ src/main/java/com/blamejared/mcbot/MCBot.java
@@ -39,41 +39,46 @@ public class MCBot {
     
     @EventSubscriber
     public void onMessageDeleted(MessageDeleteEvent event) {
-        if(event.getChannel().getName().equalsIgnoreCase("bot-log")) {
-            return;
+        if(event.getGuild().getName().equals("Modders Corner")) {
+            if(event.getChannel().getName().equalsIgnoreCase("bot-log")) {
+                return;
+            }
+            getChannel("bot-log").sendMessage(event.getAuthor().getName() + " Deleted message : ```" + event.getMessage().getContent().replaceAll("```", "") + "``` from channel: " + event.getChannel().getName());
         }
-        getChannel("bot-log").sendMessage(event.getAuthor().getName() + " Deleted message : ```" + event.getMessage().getContent().replaceAll("```", "") + "``` from channel: " + event.getChannel().getName());
     }
     
     @EventSubscriber
     public void onMessageEdited(MessageUpdateEvent event) {
-        if(event.getChannel().getName().equalsIgnoreCase("bot-log")) {
-            return;
+        if(event.getGuild().getName().equals("Modders Corner")) {
+            if(event.getChannel().getName().equalsIgnoreCase("bot-log")) {
+                return;
+            }
+            if(event.getAuthor().isBot()) {
+                return;
+            }
+            getChannel("bot-log").sendMessage(event.getAuthor().getName() + " Edited message: ```" + event.getOldMessage().getContent().replaceAll("```", "") + "``` -> ```" + event.getNewMessage().getContent().replaceAll("```", "") + "``` from channel: " + event.getChannel().getName());
         }
-        if(event.getAuthor().isBot()) {
-            return;
-        }
-        getChannel("bot-log").sendMessage(event.getAuthor().getName() + " Edited message: ```" + event.getOldMessage().getContent().replaceAll("```", "") + "``` -> ```" + event.getNewMessage().getContent().replaceAll("```", "") + "``` from channel: " + event.getChannel().getName());
-        
     }
     
     @EventSubscriber
     public void onMessageRecieved(MessageReceivedEvent event) {
-        if(event.getMessage().getChannel().getName().equals("general-discussion")) {
-            boolean isGif = false;
-            if(event.getMessage().getContent().contains(".gif")) {
-                isGif = true;
-            }
-            if(event.getMessage().getContent().contains("https://giphy.com")) {
-                isGif = true;
-            }
-            if(event.getMessage().getContent().contains("https://tenor.co")) {
-                isGif = true;
-            }
-            
-            if(isGif) {
-                event.getMessage().delete();
-                event.getChannel().sendMessage("Sorry! GIFs are not allowed in this chat! Head to <#235949539138338816>");
+        if(event.getGuild().getName().equals("Modders Corner")) {
+            if(event.getMessage().getChannel().getName().equals("general-discussion")) {
+                boolean isGif = false;
+                if(event.getMessage().getContent().contains(".gif")) {
+                    isGif = true;
+                }
+                if(event.getMessage().getContent().contains("https://giphy.com")) {
+                    isGif = true;
+                }
+                if(event.getMessage().getContent().contains("https://tenor.co")) {
+                    isGif = true;
+                }
+        
+                if(isGif) {
+                    event.getMessage().delete();
+                    event.getChannel().sendMessage("Sorry! GIFs are not allowed in this chat! Head to <#235949539138338816>");
+                }
             }
         }
     }
